% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npfixedcomp2.R
\name{estpi0}
\alias{estpi0}
\alias{estpi0.npnormll}
\alias{estpi0.npnormcvm}
\alias{estpi0.nptll}
\alias{estpi0.npnormllw}
\alias{estpi0.npnormcvmw}
\title{Computing non-parametric mixing distribution with estimated proportion at 0}
\usage{
estpi0(v, method = "npnormll", ...)

estpi0.npnormll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvm(v, beta, val, order, mix = NULL, ...)

estpi0.nptll(v, beta, val, order, mix = NULL, ...)

estpi0.npnormllw(v, beta, val, order, mix = NULL, ...)

estpi0.npnormcvmw(v, beta, val, order, mix = NULL, ...)
}
\arguments{
\item{v}{observations}

\item{method}{An implemented family; see details}

\item{...}{parameters above passed to the specific method.}

\item{beta}{structural parameter}

\item{val}{thresholding function}

\item{order}{the parameter for the binned version.}

\item{mix}{initial mixing distribution.}
}
\description{
computing non-parametric mixing distribution with estimated proportion at 0
}
\details{
This is a function for computing non-parametric mixing
distribution with estimated proportion at 0. Different families will
have different threshold values.

The parameters are listed as follows:
\itemize{
\item tol: tolerance to stop the code.
\item verbose: logical; Whether to print the intermediate results.
}

It is not shown in the parameter section since various method have different
default threshold values and this function essentially calls the class method
in the object.

The full list of implemented families is in \code{\link{computemixdist}}.
}
\examples{
data = rnorm(500, c(0, 2))
system.time(r <- estpi0(data, method = "npnormll", val = 2, verbose = TRUE))
r
system.time(r <- estpi0(data, method = "npnormllw", val = 2, verbose = TRUE, order = -2))
r
system.time(r <- estpi0(data, method = "npnormcvm", val = 0.1, verbose = TRUE))
r
system.time(r <- estpi0(data, method = "npnormcvmw", val = 0.1, verbose = TRUE, order = -2))
r
system.time(r <- estpi0(data, method = "nptll", val = 2, verbose = TRUE))
r
}
