% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnpdisct}
\alias{dnpdisct}
\title{non-parametric discrete non-central t distribution}
\usage{
dnpdisct(x, mu0, pi0, df, h, lg = FALSE)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{mu0}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points}

\item{df}{the degree of freedom}

\item{h}{the bin width}

\item{lg}{logical; if TRUE, the result will be given in log scale.}
}
\description{
The density of non-parametric discrete non-central t distribution
}
\details{
\code{dnpdisct} gives the density .

In this implementation, the support space is ..., -h, 0, h, ...
}
