% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnpnormc}
\alias{dnpnormc}
\title{non-parametric one-parameter normal distribution}
\usage{
dnpnormc(x, mu0, pi0, n, lg = FALSE)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{mu0}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points}

\item{n}{number of observations (not the length of x)}

\item{lg}{logical; if TRUE, the result will be given in log scale.}
}
\description{
The density and the distribution function of non-parametric one-parameter normal distribution
}
\details{
\code{dnpnormc} gives the density
}
