% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnppois}
\alias{dnppois}
\alias{pnppois}
\title{non-parametric normal distribution}
\usage{
dnppois(x, mu0, pi0, stdev = 1, lg = FALSE)

pnppois(x, mu0, pi0, stdev = 1, lt = TRUE, lg = FALSE)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{mu0}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points}

\item{stdev}{structure parameter, passed but not referenced, for compatibility}

\item{lg}{logical; if TRUE, the result will be given in log scale.}

\item{lt}{logical; if TRUE, the lower probability is computed}
}
\description{
The density and the distribution function of non-parametric poisson distribution
}
\details{
\code{dnppois} gives the density, \code{pnppois} gives the distribution function.

It is important to note that \code{pnppois} and its underlying function did not use
logspaceadd type implementation, the accuracy may be affected.
}
