% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnpt}
\alias{dnpt}
\title{non-parametric t distribution}
\usage{
dnpt(x, mu0, pi0, df, lg = FALSE)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{mu0}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points}

\item{df}{degree of freedom.}

\item{lg}{logical; if TRUE, the result will be given in log scale.}
}
\description{
The density and the distribution function of non-parametric t distribution
}
\details{
\code{dnpnorm} gives the density.
}
