% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnpnorm}
\alias{dnpnorm}
\alias{pnpnorm}
\title{non-parametric normal distribution}
\usage{
dnpnorm(x, mu0, pi0, stdev = 1, lg = FALSE)

pnpnorm(x, mu0, pi0, stdev = 1, lt = TRUE, lg = FALSE)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{mu0}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points}

\item{stdev}{standard deviation.}

\item{lg}{logical; if TRUE, the result will be given in log scale.}

\item{lt}{logical; if TRUE, the lower probability is computed}
}
\description{
The density and the distribution function of non-parametric normal distribution
}
\details{
\code{dnpnorm} gives the density, \code{pnpnorm} gives the distribution function.
}
