% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnpt}
\alias{dnpt}
\alias{pnpt}
\title{non-parametric t distribution}
\usage{
dnpt(x, mu0, pi0, df, lg = FALSE)

pnpt(x, mu0, pi0, df, lt = TRUE, lg = FALSE)
}
\arguments{
\item{x}{vector of observations, vector of quantiles}

\item{mu0}{the vector of support points}

\item{pi0}{the vector of weights correponding to the support points}

\item{df}{degree of freedom.}

\item{lg}{logical; if TRUE, the result will be given in log scale.}

\item{lt}{logical; if TRUE, the lower probability is computed}
}
\description{
The density and the distribution function of non-parametric t distribution
}
\details{
\code{dnpt} gives the density, \code{pnpt} gives the distribution function.
}
